requirements.txt
asgiref
certifi
charset-normalizer
decorator
Django
django-cors-headers
django-storages
djangorestframework
idna
imageio
imageio-ffmpeg
moviepy
numpy
Pillow
proglog
psycopg2-binary
pytz
requests
sqlparse
tqdm
typing_extensions
urllib3
gunicorn
